Fixes y tests de Service completado

+++ src/main/java/ISPP/G5/INDVELOPERS/controllers/ReviewController.java
@@ -4,6 +4,12 @@ package ISPP.G5.INDVELOPERS.controllers;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,7 +20,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import ISPP.G5.INDVELOPERS.models.Developer;
+import ISPP.G5.INDVELOPERS.models.Game;
 import ISPP.G5.INDVELOPERS.models.Review;
+import ISPP.G5.INDVELOPERS.services.DeveloperService;
+import ISPP.G5.INDVELOPERS.services.GameService;
 import ISPP.G5.INDVELOPERS.services.ReviewService;
 
 @CrossOrigin("*")
@@ -23,32 +33,73 @@ import ISPP.G5.INDVELOPERS.services.ReviewService;
 public class ReviewController {
 
 	@Autowired
-	private ReviewService service;
+	private ReviewService		service;
+
+	@Autowired
+	private DeveloperService	developerService;
+
+	@Autowired
+	private GameService			gameService;
 
 
-	@GetMapping("/communicate")
-	public String communicateWithReact() {
-		return service.communicateWithReact();
-	}
-	@GetMapping
-	public List<Review> findAll() {
-		return service.findAll();
+	@GetMapping("/game/{gameId}")
+	public ResponseEntity<List<Review>> findAllByGame(@PathVariable("gameId") final String gameId) {
+		try {
+			return ResponseEntity.ok(service.findAllByGameId(gameId));
+		} catch (IllegalArgumentException e) {
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+		}
 	}
 	@GetMapping("/{id}")
-	public Review findById(@PathVariable final String id) {
-		return service.findById(id);
+	public ResponseEntity<Review> findById(@PathVariable("id") final String id) throws NotFoundException {
+		try {
+			return ResponseEntity.ok(service.findById(id));
+		} catch (IllegalArgumentException e) {
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+		}
 	}
-	@PostMapping("/add")
-	public String addReview(@RequestBody final Review review) {
-		return service.addReview(review);
+	@PostMapping("/game/{gameId}/add")
+	public ResponseEntity<String> addReview(@RequestBody final Review review, @PathVariable("gameId") final String gameId) throws NotFoundException {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+		Developer developer = developerService.findByUsername(userDetails.getUsername());
+		Game game = gameService.findById(gameId);
+		try {
+			return ResponseEntity.status(HttpStatus.CREATED).body(service.addReview(review, game, developer));
+		} catch (IllegalArgumentException e) {
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+		}
 	}
 	@PutMapping("/{id}/edit")
-	public String editReview(@RequestBody final Review review, @PathVariable final String id) {
-		review.setId(id);
-		return service.updateReview(review);
+	public ResponseEntity<String> editReview(@RequestBody final Review review, @PathVariable("id") final String id) throws NotFoundException {
+		Review r = service.findById(id);
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+		Developer developer = developerService.findByUsername(userDetails.getUsername());
+		if (!review.getDeveloper().getId().equals(developer.getId()))
+			throw new IllegalArgumentException("Only the creator of the review can edit it");
+		try {
+			r.setDeveloper(review.getDeveloper());
+			r.setGame(review.getGame());
+			r.setScore(review.getScore());
+			r.setText(review.getText());
+			return new ResponseEntity<>(service.updateReview(r), HttpStatus.OK);
+		} catch (IllegalArgumentException e) {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
 	}
 	@DeleteMapping("/{id}/delete")
-	public String deleteReview(@PathVariable final String id) {
-		return service.deleteReview(id);
+	public ResponseEntity<HttpStatus> deleteReview(@PathVariable final String id) throws NotFoundException {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+		Developer developer = developerService.findByUsername(userDetails.getUsername());
+		if (!service.findById(id).getDeveloper().getId().equals(developer.getId()))
+			throw new IllegalArgumentException("Only the creator of the review can remove it");
+		try {
+			service.deleteReview(id);
+			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+		} catch (IllegalArgumentException e) {
+			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 }
